% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stadistics_selection.R
\name{calculate_median_statistics}
\alias{calculate_median_statistics}
\title{Calculate statistics for selected files}
\usage{
calculate_median_statistics(nsDirectory = NULL)
}
\arguments{
\item{nsDirectory}{The directory where the files are located. Default is NULL.}
}
\value{
A data frame containing the calculated statistics for each file.
}
\description{
This function reads in a set of files that match the pattern "^N0" and calculates
various statistics based on the data in each file. The statistics include the median,
mean, variance, minimum, maximum, number of unique values, total number of rows, and
percentage of rows per unique value for a specific variable in each file.
}
