% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stadistics_selection.R
\name{calculate_overall_statistics}
\alias{calculate_overall_statistics}
\title{Calculate statistics and return metrics dataframe}
\usage{
calculate_overall_statistics(overallDirectory = NULL, Nrun = 48)
}
\arguments{
\item{overallDirectory}{The directory where the statistics files are located.}

\item{Nrun}{The number of runs.}
}
\value{
A dataframe containing the calculated statistics.
}
\description{
This function calculates statistics from multiple files and returns a dataframe.
}
