% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_phenotypes}
\alias{clean_phenotypes}
\title{Clean phenotypes}
\usage{
clean_phenotypes(GeneDiscoveRobject = NULL)
}
\arguments{
\item{GeneDiscoveRobject}{A GeneDiscoveR object.}
}
\value{
The GeneDiscoveR object with the phenotypes removed.
}
\description{
This function removes the phenotypes from a GeneDiscoveR object.
}
\examples{
# Create a GeneDiscoveR object
N0sDir <- system.file("extdata", "N0-1dot3-6", package = "GeneDiscoveR")
overallsDir <- system.file("extdata", "Comparatives-1dot3-6", package = "GeneDiscoveR")
dataFile <- system.file("extdata", "annotatedCDSs.tsv", package = "GeneDiscoveR")
minInflation <- 1.3
maxInflation <- 6
stepInflation <- 0.1

GeneDiscoveRobject <- GeneDiscoveR(overallsDir = overallsDir, N0sDir = N0sDir, dataFile = dataFile, minInflation = minInflation, maxInflation = maxInflation, stepInflation = stepInflation)

# Set active run
GeneDiscoveRobject <- set_run_active(GeneDiscoveRobject, InflationValue = 1.8)

# Select species by phenotype
GeneDiscoveRobject <- select_species_by_phenotype(GeneDiscoveRobject = GeneDiscoveRobject, columnPhenotype = "Oil-body-type", columnID = "OrthofinderID", type = "one_in_specialized_cell")
GeneDiscoveRobject <- select_species_by_phenotype(GeneDiscoveRobject = GeneDiscoveRobject, columnPhenotype = "Oil-body-type", columnID = "OrthofinderID", type = "many_in_all_cells")

# Phenotypes
names(GeneDiscoveRobject$Phenotypes)
# Output: [1] "one_in_specialized_cell" "many_in_all_cells"

# Clean phenotypes
GeneDiscoveRobject <- clean_phenotypes(GeneDiscoveRobject)
names(GeneDiscoveRobject$Phenotypes)
# Output: NULL
}
