% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inparanoidb.R
\name{.check_OrtoA_in_OrtoB}
\alias{.check_OrtoA_in_OrtoB}
\title{Check Orthologs in Ortholog Set B}
\usage{
.check_OrtoA_in_OrtoB(dfMerge, cores = 1)
}
\arguments{
\item{dfMerge}{A data frame containing the merged data from Ortholog Set A and Ortholog Set B.}

\item{cores}{The number of cores to use for parallel processing.}
}
\value{
A list of unique indices of orthologs in Ortholog Set B that match the values in Ortholog Set A.
}
\description{
This function checks for orthologs in Ortholog Set B based on the values in Ortholog Set A.
}
\examples{
df <- data.frame(OrtoA = c("A", "B", "C"), OrtoB.1 = c("A", "B", "D"), OrtoB.2 = c("B", "C", "E"))
.check_OrtoA_in_OrtoB(df, cores = 4)
# Output: [[1]]
# [1] 1 2
#
# [[2]]
# [1] 2 3
#
# [[3]]
# integer(0)
}
\keyword{internal}
