% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.fisher_per_row}
\alias{.fisher_per_row}
\title{Calculate Fisher's exact test for each row in a data frame}
\usage{
.fisher_per_row(df, predictor, response, nameColumn1, nameColumn2, nameColumn3)
}
\arguments{
\item{df}{A data frame containing the predictor and response variables}

\item{predictor}{A vector representing the predictor variable}

\item{response}{A vector representing the response variable}

\item{nameColumn1}{The name of the column in the data frame representing the first variable in the contingency table}

\item{nameColumn2}{The name of the column in the data frame representing the second variable in the contingency table}

\item{nameColumn3}{The name of the column to be added to the data frame to store the Fisher's exact test result}
}
\value{
The input data frame with an additional column containing the Fisher's exact test result for each row
}
\description{
This function calculates Fisher's exact test for each row in a data frame. It takes a data frame, a predictor variable, a response variable, and three column names as input. It performs Fisher's exact test for each row and adds the test result as a new column to the data frame.
}
\examples{
\dontrun{
df <- data.frame(predictor = c(1, 0, 1), response = c(1, 1, 0))
df <- .fisher_per_row(df, df$predictor, df$response, "nameColumn1", "nameColumn2", "nameColumn3")
print(df)
}
}
