% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.GeneDiscoveRFilteredGenes}
\alias{.GeneDiscoveRFilteredGenes}
\title{Create a GeneDiscoveRFilteredGenes object}
\usage{
.GeneDiscoveRFilteredGenes(
  table = NULL,
  name = NULL,
  pvalue = NULL,
  oddsRatio = NULL,
  sign = NULL
)
}
\arguments{
\item{table}{The table containing the filtered genes.}

\item{name}{The name of the filtered genes.}

\item{pvalue}{The p-value threshold for filtering genes (always <= pvalue filter).}

\item{oddsRatio}{The odds ratio threshold for filtering genes.}

\item{sign}{The sign of the odds ratio for filtering genes (>, >=, <, <=).}
}
\value{
A GeneDiscoveRFilteredGenes object containing the filtered genes.
}
\description{
This function creates a GeneDiscoveRFilteredGenes object, which is used to store filtered genes
based on specified criteria such as table, name, p-value, odds ratio, and sign.
}
\examples{
# Create a GeneDiscoveRFilteredGenes object
\dontrun{
filtered_genes <- .GeneDiscoveRFilteredGenes(table = my_table, name = "Filtered Genes", pvalue = 0.05, oddsRatio = 2, sign = ">")
}
}
