% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.GeneDiscoveRIdentification}
\alias{.GeneDiscoveRIdentification}
\title{GeneDiscoveRIdentification Function}
\usage{
.GeneDiscoveRIdentification(
  name = NULL,
  statistic = NULL,
  formula = NULL,
  Phenotypes = NULL,
  columns = NULL
)
}
\arguments{
\item{name}{A character string specifying the name of the object.}

\item{statistic}{A character string specifying the statistic to be used.}

\item{formula}{A formula specifying the model formula.}

\item{Phenotypes}{A character vector specifying the phenotypes.}

\item{columns}{A character vector specifying the columns to be used.}
}
\value{
A GeneDiscoveRIdentification object with the specified parameters.
}
\description{
This function creates a GeneDiscoveRIdentification object with the specified parameters.
}
\examples{
\dontrun{
GeneDiscoveRIdentification(name = "MyIdentification", statistic = "Fisher", formula = as.formula("A ~ B"), Phenotypes = c("A", "B"), columns = c("spIDs", "Phenotype1"))
}
}
\keyword{internal}
