% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inparanoidb.R
\name{.get_even_values}
\alias{.get_even_values}
\title{Get the even values from a vector}
\usage{
.get_even_values(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
A numeric vector containing only the even values from the input vector
}
\description{
This function takes a vector as input and returns a new vector containing only the even values from the input vector.
}
\examples{
x <- c(1, 2, 3, 4, 5, 6)
get_even_values(x)
# Output: 2 4 6
}
