% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.index_N0_inflation}
\alias{.index_N0_inflation}
\title{Calculate the index of a specific inflation value in a GeneDiscoveR object.}
\usage{
.index_N0_inflation(GeneDiscoveRobject = NULL, InflationValue = 1.8)
}
\arguments{
\item{GeneDiscoveRobject}{A GeneDiscoveR object.}

\item{InflationValue}{The inflation value for which the index needs to be calculated.}
}
\value{
The index of the specified inflation value in the GeneDiscoveR object.
}
\description{
This function calculates the index of a specific inflation value in a GeneDiscoveR object.
}
\examples{
# Create a GeneDiscoveR object
N0sDir <- system.file("extdata", "N0-1dot3-6", package = "GeneDiscoveR")
overallsDir <- system.file("extdata", "Comparatives-1dot3-6", package = "GeneDiscoveR")
dataFile <- system.file("extdata", "annotatedCDSs.tsv", package = "GeneDiscoveR")
minInflation <- 1.3
maxInflation <- 6
stepInflation <- 0.1

GeneDiscoveRobject <- GeneDiscoveR(overallsDir = overallsDir, N0sDir = N0sDir, dataFile = dataFile, minInflation = minInflation, maxInflation = maxInflation, stepInflation = stepInflation)

# Set active run
GeneDiscoveRobject <- set_run_active(GeneDiscoveRobject, InflationValue = 1.8)

# Calculate the index of inflation value 1.8
index <- .index_N0_inflation(GeneDiscoveRobject, 1.8)
index
# Output: 6
}
\keyword{internal}
