% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{.launch_genediscover_web_app}
\alias{.launch_genediscover_web_app}
\title{Launches the GeneDiscoveR Web App}
\usage{
.launch_genediscover_web_app()
}
\description{
This function creates and launches the GeneDiscoveR web application. The web app provides a user interface for analyzing gene data using the GeneDiscoveR package.
}
\examples{
\dontrun{
.launch_genediscover_web_app()
}
}
\keyword{internal}
