% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_annotation.R
\name{.map_gene_annotation}
\alias{.map_gene_annotation}
\title{Maps gene annotations based on a GeneDiscoveR object and filtered genes index.}
\usage{
.map_gene_annotation(
  GeneDiscoveRobject = NULL,
  indexFilteredGenes = NULL,
  specieWithAnnotation = "MpTAKv6-Marchantia_polymorpha_rudelaris"
)
}
\arguments{
\item{GeneDiscoveRobject}{A GeneDiscoveR object containing gene information.}

\item{indexFilteredGenes}{An index specifying the filtered genes to use for mapping annotations.}

\item{specieWithAnnotation}{The species with annotation to use for mapping.}
}
\value{
A tibble containing the mapped gene annotations.
}
\description{
This function takes a GeneDiscoveR object and an index of filtered genes as input,
and maps gene annotations based on the specified species with annotation.
}
\examples{
\dontrun{
# Map gene annotations for filtered genes at index 1
mapped_annotations <- .map_gene_annotation(GeneDiscoveRobject, indexFilteredGenes = 1)

# Print the mapped annotations
print(mapped_annotations)
}
}
\keyword{internal}
