% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inparanoidb.R
\name{.merge_groups}
\alias{.merge_groups}
\title{Merge groups based on given results}
\usage{
.merge_groups(dfMerge, resultados)
}
\arguments{
\item{dfMerge}{A data frame containing the data to be merged}

\item{resultados}{A list of results used for merging groups}
}
\value{
A list containing two data frames: \code{dfMergemodifyunprocess} and
\code{dfMergemodifyprocess}
}
\description{
This function merges groups based on the given results. It takes a data frame
\code{dfMerge} and a list \code{resultados} as input. It returns a list
containing two data frames: \code{dfMergemodifyunprocess} and
\code{dfMergemodifyprocess}.
}
\examples{
\dontrun{
df <- data.frame(
    OrtoA = c("A", "B", "C"),
    OrtoBsplit1 = c("1,2,3", "4,5,6", "7,8,9"),
    OrtoBsplit2 = c("10,11,12", "13,14,15", "16,17,18")
)
resultados <- list(c("1\t2\t3", "4\t5\t6"), c("7\t8\t9"))
.merge_groups(df, resultados)
}
}
