% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inparanoidb.R
\name{.merge_groups_parallel}
\alias{.merge_groups_parallel}
\title{Merge groups in parallel}
\usage{
.merge_groups_parallel(dfMergemodify, cores = 1)
}
\arguments{
\item{dfMergemodify}{The dataframe to be modified and merged.}

\item{cores}{The number of cores to be used for parallel processing. Default is 1.}
}
\value{
The modified and merged dataframe.
}
\description{
This function merges groups in parallel using the provided dataframe.
}
