% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inparanoidb.R
\name{.merge_inParanoid_files}
\alias{.merge_inParanoid_files}
\title{Merge inParanoid Files}
\usage{
.merge_inParanoid_files(directory)
}
\arguments{
\item{directory}{The directory where the inParanoid files are located.}
}
\value{
A merged dataframe containing the data from all the inParanoid files.
}
\description{
This function merges multiple inParanoid files into a single dataframe.
}
\examples{
# Set the directory where the inParanoid files are located
directory <- system.file("extdata", "inparanoidbFiles", package = "GeneDiscoveR")

# Merge the inParanoid files
merged_df <- .merge_inParanoid_files(directory)

# Print the merged dataframe
\dontrun{
print(merged_df)
}

}
