% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{.server_genediscover}
\alias{.server_genediscover}
\title{Server function for GeneDiscoveR}
\usage{
.server_genediscover(input, output, session)
}
\arguments{
\item{input}{The input values from the UI.}

\item{output}{The output objects to be rendered in the UI.}

\item{session}{The session object for the Shiny application.}
}
\value{
None
}
\description{
This function defines the server logic for the GeneDiscoveR application.
It handles the rendering of the menu, the selection of HOGs in the volcano plot,
and the filtering of data based on selected genes.
}
\examples{
\dontrun{
.server_genediscover(input, output, session)
}
}
