% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_visualization_export.R
\name{.set_ggplot2_theme}
\alias{.set_ggplot2_theme}
\title{Set ggplot2 theme}
\usage{
.set_ggplot2_theme()
}
\value{
This function does not return any value.
}
\description{
This function sets a custom theme for ggplot2 plots.
}
\details{
The function sets various theme elements for ggplot2 plots, including the legend background,
plot title and subtitle, panel grid and background, axis text and title, legend key, and panel border.
The function also sets the theme using \code{theme_set()} function from ggplot2 package.
}
\examples{
.set_ggplot2_theme()
}
\keyword{internal}
