% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.sum_per_phenotype}
\alias{.sum_per_phenotype}
\title{Calculate the sum per phenotype}
\usage{
.sum_per_phenotype(df, predictor, response, nameColumn1, nameColumn2)
}
\arguments{
\item{df}{The input data frame.}

\item{predictor}{The name of the predictor variable.}

\item{response}{The name of the response variable.}

\item{nameColumn1}{The name of the column to store the sum for the predictor variable.}

\item{nameColumn2}{The name of the column to store the sum for the response variable.}
}
\value{
The modified data frame with additional columns storing the sums per phenotype.
}
\description{
This function calculates the sum of non-missing values per phenotype in a data frame.
}
\examples{
\dontrun{
df <- data.frame(predictor = c(1, 2, NA, 4), response = c(NA, 2, 3, 4))
.sum_per_phenotype(df, "predictor", "response", "sum_predictor", "sum_response")
}
}
\keyword{internal}
