% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.table_as_character}
\alias{.table_as_character}
\title{Convert a table to character vector}
\usage{
.table_as_character(table = NULL)
}
\arguments{
\item{table}{The table to be converted. It should be a data frame or a tibble.}
}
\value{
A character vector representing the table.
}
\description{
This function takes a table as input and converts it to a character vector.
}
\examples{
table <- data.frame(x = c(1, 2, 3), y = c("a", "b", "c"))
.table_as_character(table)
# Output: [1] "1" "2" "3" "a" "b" "c"
}
\keyword{internal}
