% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inparanoidb.R
\name{.update_dfMerge}
\alias{.update_dfMerge}
\title{Update the dfMerge dataframe with orthologous gene IDs from InParanoid database}
\usage{
.update_dfMerge(dfMerge, at_prefix, datafile, columnID)
}
\arguments{
\item{dfMerge}{The input dataframe to be updated.}

\item{at_prefix}{The prefix used in the gene IDs to be replaced.}

\item{datafile}{The datafile containing the orthologous gene IDs from the InParanoid database.}

\item{columnID}{The column name in the datafile that contains the orthologous gene IDs.}
}
\value{
The updated dfMerge dataframe with cleaned orthologous gene IDs.
}
\description{
This function takes a dataframe (dfMerge) and updates it by replacing the orthologous gene IDs
in the columns starting with "OrtoBsplit" with cleaned IDs obtained from the InParanoid database.
The function requires the following parameters:
}
\details{
\code{dfMerge}: The input dataframe to be updated.

\code{at_prefix}: The prefix used in the gene IDs to be replaced.

\code{datafile}: The datafile containing the orthologous gene IDs from the InParanoid database.

\code{columnID}: The column name in the datafile that contains the orthologous gene IDs.

The function iterates over each row of the dfMerge dataframe and for each row, it extracts the
orthologous gene IDs from the columns starting with "OrtoBsplit". It then cleans the IDs by removing
the prefix and any additional characters. The cleaned IDs are then replaced in the dfMerge dataframe.
Additionally, the function performs some data cleaning operations on the dfMerge dataframe, such as
removing leading and trailing whitespaces and removing consecutive commas in the columns starting with "OrtoBsplit".
Finally, the function renames the columns in the dfMerge dataframe based on the orthologous gene IDs
obtained from the datafile.
}
\examples{
# Example usage of the update_dfMerge function
\dontrun{
df <- .update_dfMerge(dfMerge, "AT", datafile, "geneID")
}
}
