% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{GeneDiscoveR}
\alias{GeneDiscoveR}
\title{Create a GeneDiscoveR object}
\usage{
GeneDiscoveR(
  overallsDir = NULL,
  N0sDir = NULL,
  dataFile = NULL,
  annotationFile = NULL,
  uniqueInflation = NULL,
  minInflation = NULL,
  maxInflation = NULL,
  stepInflation = NULL,
  pairSpeciesDir = NULL,
  principalSpecie = NULL,
  principalSpeciePrefix = NULL,
  orthologsTool = "OrthoFinder"
)
}
\arguments{
\item{overallsDir}{The directory path for the overalls (only for OrthoFinder run).}

\item{N0sDir}{The directory path for the N0s (only for OrthoFinder run).}

\item{dataFile}{The file path for the table with at least two columns (speciesID and n columns one per trait or phenotype) data.}

\item{annotationFile}{The file path for the annotation data.}

\item{uniqueInflation}{The unique inflation value to be used (only for OrthoFinder run).}

\item{minInflation}{The minimum inflation value for the range of inflations (only for OrthoFinder run).}

\item{maxInflation}{The maximum inflation value for the range of inflations (only for OrthoFinder run).}

\item{stepInflation}{The step size for the range of inflations (only for OrthoFinder run).}

\item{pairSpeciesDir}{The directory path for the pair species (only for InParanoiDB run).}

\item{principalSpecie}{The principal species (only for InParanoiDB run).}

\item{principalSpeciePrefix}{The principal species prefix (only for InParanoiDB run).}

\item{orthologsTool}{The tool used for orthologs identification, default is OrthoFinder (OrthoFinder or InParanoiDB).}
}
\value{
A GeneDiscoveR object with the specified parameters.
}
\description{
This function creates a GeneDiscoveR object with the specified parameters.
The data can be originated from OrthoFinder or InParanoiDB. Can be indicate the origin of the orthologs definition with the parameter \code{orthologsTool}.
}
\details{
For OrthoFinder, the parameters \code{overallsDir}, \code{N0sDir}, \code{dataFile}, \code{uniqueInflation}, \code{minInflation}, \code{maxInflation}, and \code{stepInflation} are required.
The parameters of Inflation depend on the type of analysis to be performed with unique inflation or a range of inflations.

For InParanoiDB, the parameters \code{pairSpeciesDir}, \code{principalSpecie}, and \code{principalSpeciePrefix} are required.
The comparison is made between the principal species and the other species. The principal species is the species that will be used as a reference.
The principal species name is \code{principalSpecie} and the principal species gene prefix is \code{principalSpeciePrefix}.
}
\examples{
# Create a GeneDiscoveR for an execution with multiple inflation values

N0sDir <- system.file("extdata", "N0-1dot3-6", package = "GeneDiscoveR")
overallsDir <- system.file("extdata", "Comparatives-1dot3-6", package = "GeneDiscoveR")
dataFile <- system.file("extdata", "annotatedCDSs.tsv", package = "GeneDiscoveR")
minInflation <- 1.3
maxInflation <- 6
stepInflation <- 0.1

GeneDiscoveRobject <- GeneDiscoveR(overallsDir = overallsDir, N0sDir = N0sDir, dataFile = dataFile, minInflation = minInflation, maxInflation = maxInflation, stepInflation = stepInflation)

# Create a GeneDiscoveR for an execution with only one inflation value

uniqueInflation <- 1.8

GeneDiscoveRobject <- GeneDiscoveR(N0sDir = N0sDir, dataFile = dataFile, uniqueInflation = uniqueInflation)

# If you want to indicate the annotation file, you can use the following code:

annotationFile <- system.file("extdata", "annotatedCDSs.tsv")

GeneDiscoveRobject <- GeneDiscoveR(overallsDir = overallsDir, N0sDir = N0sDir, annotationFile = annotationFile, dataFile = dataFile, minInflation = minInflation, maxInflation = maxInflation, stepInflation = stepInflation)
}
