% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_complete_table}
\alias{get_complete_table}
\title{Get complete table from GeneDiscoveR object}
\usage{
get_complete_table(GeneDiscoveRobject = NULL)
}
\arguments{
\item{GeneDiscoveRobject}{A GeneDiscoveR object.}
}
\value{
The complete table from the GeneDiscoveR object.
}
\description{
This function retrieves the complete table from a GeneDiscoveR object.
If you perform a gene identification, the complete table will be the table with the results of the gene identification.
}
\examples{
N0sDir <- system.file("extdata", "N0-1dot3-6", package = "GeneDiscoveR")
overallsDir <- system.file("extdata", "Comparatives-1dot3-6", package = "GeneDiscoveR")
dataFile <- system.file("extdata", "annotatedCDSs.tsv", package = "GeneDiscoveR")
minInflation <- 1.3
maxInflation <- 6
stepInflation <- 0.1

GeneDiscoveRobject <- GeneDiscoveR(overallsDir = overallsDir, N0sDir = N0sDir, dataFile = dataFile, minInflation = minInflation, maxInflation = maxInflation, stepInflation = stepInflation)

# Set active run
GeneDiscoveRobject <- set_run_active(GeneDiscoveRobject, InflationValue = 1.8)

# Get complete table
table <- get_complete_table(GeneDiscoveRobject)
# Output: A data frame with the complete table from the GeneDiscoveR object.
}
