% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_visualization_export.R
\name{get_filtered_genes_table}
\alias{get_filtered_genes_table}
\title{Get Filtered Genes Table}
\usage{
get_filtered_genes_table(
  GeneDiscoveRobject = NULL,
  name = NULL,
  pvalue = NULL,
  oddsRatio = NULL,
  sign = NULL
)
}
\arguments{
\item{GeneDiscoveRobject}{A GeneDiscoveR object containing the filtered genes.}

\item{name}{The name of the gene.}

\item{pvalue}{The p-value of the gene.}

\item{oddsRatio}{The odds ratio of the gene.}

\item{sign}{The sign of the gene.}
}
\value{
A filtered genes table that matches the specified criteria.
}
\description{
This function retrieves a filtered genes table based on specified criteria.
}
\examples{
# Create a GeneDiscoveR object
N0sDir <- system.file("extdata", "N0-1dot3-6", package = "GeneDiscoveR")
overallsDir <- system.file("extdata", "Comparatives-1dot3-6", package = "GeneDiscoveR")
dataFile <- system.file("extdata", "annotatedCDSs.tsv", package = "GeneDiscoveR")
minInflation <- 1.3
maxInflation <- 6
stepInflation <- 0.1

GeneDiscoveRobject <- GeneDiscoveR(overallsDir = overallsDir, N0sDir = N0sDir, dataFile = dataFile, minInflation = minInflation, maxInflation = maxInflation, stepInflation = stepInflation)

# Set active run
GeneDiscoveRobject <- set_run_active(GeneDiscoveRobject, InflationValue = 1.8)

# Select species by phenotype
GeneDiscoveRobject <- select_species_by_phenotype(GeneDiscoveRobject = GeneDiscoveRobject, columnPhenotype = "Oil-body-type", columnID = "OrthofinderID", type = "one_in_specialized_cell")
GeneDiscoveRobject <- select_species_by_phenotype(GeneDiscoveRobject = GeneDiscoveRobject, columnPhenotype = "Oil-body-type", columnID = "OrthofinderID", type = "many_in_all_cells")

# Gene identification by phenotype
GeneDiscoveRobject <- gene_identification_by_phenotype(GeneDiscoveRobject = GeneDiscoveRobject, formula = as.formula("one_in_specialized_cell ~ many_in_all_cells"), statistic = "Fisher", name = "PerOBtype", cores = 8)

# Select genes by phenotype
GeneDiscoveRobject <- select_genes_by_phenotype(GeneDiscoveRobject, pvalue = 0.05, oddsRatio = 1, sign = ">=", name = "PerOBtype")

# Get filtered genes table
get_filtered_genes_table(GeneDiscoveRobject, name = "PerOBtype", pvalue = 0.05, oddsRatio = 1, sign = ">=")
}
