% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_visualization_export.R
\name{get_medians_statistics}
\alias{get_medians_statistics}
\title{Get the median statistics from a GeneDiscoveR object}
\usage{
get_medians_statistics(GeneDiscoveRobject = NULL)
}
\arguments{
\item{GeneDiscoveRobject}{A GeneDiscoveR object.}
}
\value{
A numeric vector containing the median statistics.
}
\description{
This function retrieves the median statistics from a GeneDiscoveR object.
}
\examples{
# Create a GeneDiscoveR object
N0sDir <- system.file("extdata", "N0-1dot3-6", package = "GeneDiscoveR")
overallsDir <- system.file("extdata", "Comparatives-1dot3-6", package = "GeneDiscoveR")
dataFile <- system.file("extdata", "annotatedCDSs.tsv", package = "GeneDiscoveR")
minInflation <- 1.3
maxInflation <- 6
stepInflation <- 0.1

GeneDiscoveRobject <- GeneDiscoveR(overallsDir = overallsDir, N0sDir = N0sDir, dataFile = dataFile, minInflation = minInflation, maxInflation = maxInflation, stepInflation = stepInflation)

# Calculate median statistics
GeneDiscoveRobject <- calculate_median_statistics(GeneDiscoveRobject, cores = 8)

# Get median statistics
medianStatistics <- get_medians_statistics(GeneDiscoveRobject)
# Output: A data frame containing the median statistics from the GeneDiscoveR object.
}
\seealso{
\code{\link{GeneDiscoveR}}
\code{\link{calculate_median_statistics}}
\code{\link{plot_OGs_HOGs_per_inflation}}
}
