% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_names_identification}
\alias{get_names_identification}
\title{Get the names of the identification objects in a GeneDiscoveR object}
\usage{
get_names_identification(GeneDiscoveRobject = NULL)
}
\arguments{
\item{GeneDiscoveRobject}{A GeneDiscoveR object.}
}
\value{
A character vector containing the names of the identification objects.
}
\description{
This function takes a GeneDiscoveR object as input and returns a vector
containing the names of the identification objects within the GeneDiscoveR
object.
}
\examples{
# Create a GeneDiscoveR object
N0sDir <- system.file("extdata", "N0-1dot3-6", package = "GeneDiscoveR")
overallsDir <- system.file("extdata", "Comparatives-1dot3-6", package = "GeneDiscoveR")
dataFile <- system.file("extdata", "annotatedCDSs.tsv", package = "GeneDiscoveR")
minInflation <- 1.3
maxInflation <- 6
stepInflation <- 0.1

GeneDiscoveRobject <- GeneDiscoveR(overallsDir = overallsDir, N0sDir = N0sDir, dataFile = dataFile, minInflation = minInflation, maxInflation = maxInflation, stepInflation = stepInflation)

# Set active run
GeneDiscoveRobject <- set_run_active(GeneDiscoveRobject, InflationValue = 1.8)

# Select species by phenotype
GeneDiscoveRobject <- select_species_by_phenotype(GeneDiscoveRobject = GeneDiscoveRobject, columnPhenotype = "Oil-body-type", columnID = "OrthofinderID", type = "one_in_specialized_cell")
GeneDiscoveRobject <- select_species_by_phenotype(GeneDiscoveRobject = GeneDiscoveRobject, columnPhenotype = "Oil-body-type", columnID = "OrthofinderID", type = "many_in_all_cells")

# Gene identification by phenotype
GeneDiscoveRobject <- gene_identification_by_phenotype(GeneDiscoveRobject = GeneDiscoveRobject, formula = as.formula("one_in_specialized_cell ~ many_in_all_cells"), statistic = "Fisher", name = "OneInSpeVSmanyInAll", cores = 8)
GeneDiscoveRobject <- gene_identification_by_phenotype(GeneDiscoveRobject = GeneDiscoveRobject, formula = as.formula("many_in_all_cells ~ one_in_specialized_cell"), statistic = "Fisher", name = "manyInAllVSOneInSpe", cores = 8)

# Get the names of the identification objects
get_names_identification(GeneDiscoveRobject)
# Output: [1] "OneInSpeVSmanyInAll" "manyInAllVSOneInSpe"

}
