% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_visualization_export.R
\name{get_overall_statistics}
\alias{get_overall_statistics}
\title{Get overall statistics from GeneDiscoveR object}
\usage{
get_overall_statistics(GeneDiscoveRobject = NULL)
}
\arguments{
\item{GeneDiscoveRobject}{A GeneDiscoveR object.}
}
\value{
A data frame containing the overall metrics from the GeneDiscoveR object.
}
\description{
This function retrieves the overall statistics from a GeneDiscoveR object.
}
\examples{
# Create a GeneDiscoveR object
N0sDir <- system.file("extdata", "N0-1dot3-6", package = "GeneDiscoveR")
overallsDir <- system.file("extdata", "Comparatives-1dot3-6", package = "GeneDiscoveR")
dataFile <- system.file("extdata", "annotatedCDSs.tsv", package = "GeneDiscoveR")
minInflation <- 1.3
maxInflation <- 6
stepInflation <- 0.1

GeneDiscoveRobject <- GeneDiscoveR(overallsDir = overallsDir, N0sDir = N0sDir, dataFile = dataFile, minInflation = minInflation, maxInflation = maxInflation, stepInflation = stepInflation)

# Calculate overall statistics

GeneDiscoveRobject <- calculate_overall_statistics(GeneDiscoveRobject, cores = 8)

# Get overall statistics
overallStatistics <- get_overall_statistics(GeneDiscoveRobject)
# Output: A data frame containing the overall metrics from the GeneDiscoveR object.

}
\seealso{
\code{\link{GeneDiscoveR}}
\code{\link{calculate_overall_statistics}}
\code{\link{plot_allSpeciesOGs_sOGs_per_inflation}}
}
