% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stadistics_selection.R
\name{plot_allSpeciesOGs_sOGs_per_inflation}
\alias{plot_allSpeciesOGs_sOGs_per_inflation}
\title{Plot Metrics of Orthogroups}
\usage{
plot_allSpeciesOGs_sOGs_per_inflation(metrics = NULL, Nrun = 48)
}
\arguments{
\item{metrics}{A dataframe containing the calculated statistics.}

\item{Nrun}{The number of runs. Default is 48.}
}
\value{
An object of class "ggplot" representing the plot.
}
\description{
This function plots the metrics of orthogroups using the provided dataframe.
}
