% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_visualization_export.R
\name{plot_genediscover_detector_volcano}
\alias{plot_genediscover_detector_volcano}
\title{Plot Gene Detector Volcano}
\usage{
plot_genediscover_detector_volcano(
  GeneDiscoveRobject = NULL,
  annotationTable = NULL,
  name = NULL,
  title = "",
  type = "TPStype",
  categories = c("Diterpene", "Bacterial", "Fungi")
)
}
\arguments{
\item{GeneDiscoveRobject}{The GeneDiscoveR object containing the gene detection results.}

\item{annotationTable}{The annotation table used for filtering genes based on categories. With OG, Gene Tree Parent Clade, GeneID and type columns.}

\item{name}{The name of the gene detection result to plot.}

\item{title}{The title of the volcano plot.}

\item{type}{The type of genes to plot (default is "TPStype"). Is the column name in the annotationTable.}

\item{categories}{The categories of genes to include in the plot (default is c("Diterpene", "Bacterial", "Fungi")).}
}
\value{
A ggplot object representing the volcano plot.
}
\description{
This function plots a volcano plot for gene detection using GeneDiscoveR.
}
\examples{
# Create a GeneDiscoveR object
N0sDir <- system.file("extdata", "N0-1dot3-6", package = "GeneDiscoveR")
overallsDir <- system.file("extdata", "Comparatives-1dot3-6", package = "GeneDiscoveR")
dataFile <- system.file("extdata", "annotatedCDSs.tsv", package = "GeneDiscoveR")
minInflation <- 1.3
maxInflation <- 6
stepInflation <- 0.1

GeneDiscoveRobject <- GeneDiscoveR(overallsDir = overallsDir, N0sDir = N0sDir, dataFile = dataFile, minInflation = minInflation, maxInflation = maxInflation, stepInflation = stepInflation)

# Set active run
GeneDiscoveRobject <- set_run_active(GeneDiscoveRobject, InflationValue = 1.8)

# Select species by phenotype
GeneDiscoveRobject <- select_species_by_phenotype(GeneDiscoveRobject = GeneDiscoveRobject, columnPhenotype = "Oil-body-type", columnID = "OrthofinderID", type = "one_in_specialized_cell")
GeneDiscoveRobject <- select_species_by_phenotype(GeneDiscoveRobject = GeneDiscoveRobject, columnPhenotype = "Oil-body-type", columnID = "OrthofinderID", type = "many_in_all_cells")

# Gene identification by phenotype
GeneDiscoveRobject <- gene_identification_by_phenotype(GeneDiscoveRobject = GeneDiscoveRobject, formula = as.formula("many_in_all_cells ~ one_in_specialized_cell"), statistic = "Fisher", name = "PerType", cores = 8)

# Load the TPS genes
TPSgenes <- read.table(system.file("extdata", "TPSgenes.tsv", package = "GeneDiscoveR"), header = T, sep = "\t")
TPScategories <- c("Diterpene", "Bacterial", "Fungi")
title <- "Fisher's Exact Test per Type of Oil-body with TPS genes"
plot_genediscover_detector_volcano(GeneDiscoveRobject, annotationTable = TPSgenes, title = title, name = "PerType", type = "TPStype", categories = TPScategories)
# Volcano plot with TPS genes

# Output: A ggplot object representing the volcano plot.
}
\seealso{
\code{\link{GeneDiscoveR}}
\code{\link{plot_genediscover_volcano}}
\code{\link{run_genediscover_web_app}}
}
