% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stadistics_selection.R
\name{plot_OGs_HOGs_per_inflation}
\alias{plot_OGs_HOGs_per_inflation}
\title{Plot Metrics of Orthogroups and HOGs}
\usage{
plot_OGs_HOGs_per_inflation(medians = NULL, Nrun = 48)
}
\arguments{
\item{medians}{A dataframe containing the calculated statistics.}

\item{Nrun}{The number of runs.}
}
\value{
An object of class "ggplot" representing the plot.
}
\description{
This function plots the metrics of orthogroups and HOGs using the provided dataframe.
}
