% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_visualization_export.R
\name{run_genediscover_web_app}
\alias{run_genediscover_web_app}
\title{Run the GeneDiscover web application}
\usage{
run_genediscover_web_app()
}
\value{
None
}
\description{
This function launches the GeneDiscover web application.
}
\examples{
\dontrun{
run_genediscover_web_app()
}
}
