% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_annotation.R
\name{set_annotation_file}
\alias{set_annotation_file}
\title{Set the annotation file for a GeneDiscoveR object}
\usage{
set_annotation_file(GeneDiscoveRobject = NULL, annotationFile = NULL)
}
\arguments{
\item{GeneDiscoveRobject}{A GeneDiscoveR object to which the annotation file will be set.}

\item{annotationFile}{The path to the annotation file.}
}
\value{
The updated GeneDiscoveR object with the annotation file set.
}
\description{
This function sets the annotation file for a GeneDiscoveR object.
The annotation file contains information about the genes, such as their names, IDs, and functional annotations.
The annotation file should be in tab-separated values (TSV) format, with the first column containing the gene IDs and the subsequent columns containing the annotations in one column or multiple columns.
}
\examples{
# Create a GeneDiscoveR object
N0sDir <- system.file("extdata", "N0-1dot3-6", package = "GeneDiscoveR")
overallsDir <- system.file("extdata", "Comparatives-1dot3-6", package = "GeneDiscoveR")
dataFile <- system.file("extdata", "annotatedCDSs.tsv", package = "GeneDiscoveR")
minInflation <- 1.3
maxInflation <- 6
stepInflation <- 0.1

GeneDiscoveRobject <- GeneDiscoveR(overallsDir = overallsDir, N0sDir = N0sDir, dataFile = dataFile, minInflation = minInflation, maxInflation = maxInflation, stepInflation = stepInflation)

# Set the annotation file
annotationFile <- system.file("extdata", "MpTak_v6.1_func_annotation_1line.tsv", package = "GeneDiscoveR")
GeneDiscoveRobject <- set_annotation_file(GeneDiscoveRobject, annotationFile = annotationFile)
}
