% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_run_active}
\alias{set_run_active}
\title{Set the active run for GeneDiscoveR}
\usage{
set_run_active(GeneDiscoveRobject = NULL, InflationValue = 1.8, cores = 1)
}
\arguments{
\item{GeneDiscoveRobject}{The GeneDiscoveR object for which the active run needs to be set.}

\item{InflationValue}{The inflation value to be set for the active run, only for OrthoFinder run.}

\item{cores}{The number of cores to be used for parallel processing (default is 1), only for InParanoiDB run.}
}
\value{
The updated GeneDiscoveR object with the active run set.
}
\description{
This function sets the active run for the GeneDiscoveR object by updating the relevant fields in the object.
}
\examples{
# Create a GeneDiscoveR object
N0sDir <- system.file("extdata", "N0-1dot3-6", package = "GeneDiscoveR")
overallsDir <- system.file("extdata", "Comparatives-1dot3-6", package = "GeneDiscoveR")
dataFile <- system.file("extdata", "annotatedCDSs.tsv", package = "GeneDiscoveR")
minInflation <- 1.3
maxInflation <- 6
stepInflation <- 0.1

GeneDiscoveRobject <- GeneDiscoveR(overallsDir = overallsDir, N0sDir = N0sDir, dataFile = dataFile, minInflation = minInflation, maxInflation = maxInflation, stepInflation = stepInflation)

# Set active run
GeneDiscoveRobject <- set_run_active(GeneDiscoveRobject, InflationValue = 1.8)
}
