% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_runs_statistics.R
\name{calculate_overall_statistics}
\alias{calculate_overall_statistics}
\title{Calculate statistics and return metrics dataframe}
\usage{
calculate_overall_statistics(GeneDiscoveRobject = NULL, cores = 1)
}
\arguments{
\item{GeneDiscoveRobject}{The GeneDiscoveR object containing the overall directory and number of runs.}

\item{cores}{The number of cores to use for parallel processing. Default is 1.}
}
\value{
A GeneDiscoveR object with the calculated statistics added to the runsData$overallMetrics field.
}
\description{
This function calculates obtain the overall statistics for the runs and returns the metrics dataframe.
The metrics include the mean, median, G50 (assigned genes), G50 (all genes), O50 (assigned genes),
O50 (all genes), All species OG, and sOGs per run.
}
\examples{
# Create a GeneDiscoveR object
N0sDir <- system.file("extdata", "N0-1dot3-6", package = "GeneDiscoveR")
overallsDir <- system.file("extdata", "Comparatives-1dot3-6", package = "GeneDiscoveR")
dataFile <- system.file("extdata", "annotatedCDSs.tsv", package = "GeneDiscoveR")
minInflation <- 1.3
maxInflation <- 6
stepInflation <- 0.1

GeneDiscoveRobject <- GeneDiscoveR(overallsDir = overallsDir, N0sDir = N0sDir, dataFile = dataFile, minInflation = minInflation, maxInflation = maxInflation, stepInflation = stepInflation)

# Calculate overall statistics

GeneDiscoveRobject <- calculate_overall_statistics(GeneDiscoveRobject, cores = 8)

}
