% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_runs_statistics.R
\name{plot_OGs_HOGs_per_inflation}
\alias{plot_OGs_HOGs_per_inflation}
\title{Plot OGs and HOGs per inflation}
\usage{
plot_OGs_HOGs_per_inflation(GeneDiscoveRobject = NULL)
}
\arguments{
\item{GeneDiscoveRobject}{An object of class GeneDiscoveR. Default is NULL.}
}
\value{
An object of class "ggplot" representing the plot.
}
\description{
This function plots the number of orthogroups (OGs) and hierarchical orthogroups (HOGs) per inflation.
If a run is active, a vertical dashed line is added to the plot on the active run's inflation value.
The function is only for executions with multiple inflation values.
}
\examples{
# Create a GeneDiscoveR object
N0sDir <- system.file("extdata", "N0-1dot3-6", package = "GeneDiscoveR")
overallsDir <- system.file("extdata", "Comparatives-1dot3-6", package = "GeneDiscoveR")
dataFile <- system.file("extdata", "annotatedCDSs.tsv", package = "GeneDiscoveR")
minInflation <- 1.3
maxInflation <- 6
stepInflation <- 0.1

GeneDiscoveRobject <- GeneDiscoveR(overallsDir = overallsDir, N0sDir = N0sDir, dataFile = dataFile, minInflation = minInflation, maxInflation = maxInflation, stepInflation = stepInflation)

# Calculate median statistics

GeneDiscoveRobject <- calculate_median_statistics(GeneDiscoveRobject, cores = 8)

# Plot number of orthogroups (OGs) and hierarchical orthogroups (HOGs) per inflation
plot_OGs_HOGs_per_inflation(GeneDiscoveRobject)

# Plot number of orthogroups (OGs) and hierarchical orthogroups (HOGs) per inflation with active run

GeneDiscoveRobject <- set_run_active(GeneDiscoveRobject, InflationValue = 1.8)
plot_OGs_HOGs_per_inflation(GeneDiscoveRobject)

}
